:recycle: Refactor Timer
